apps/trading-api/requirements.txt
# Enhanced Trading API Requirements
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyjwt==2.8.0
python-multipart==0.0.6
websockets==12.0
slowapi==0.1.9
structlog==23.2.0
redis==5.0.0
prometheus-client==0.19.0
psutil==5.9.0
aiofiles==23.2.0
httpx==0.25.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9